several changes (sorry Swapneel)
- Now it is possible to execute user-defined python scripts inside the
containers: instead of having `-s script system` with only a selection
of scripts, we have `script -n system`, so that if the string `script`
is a known script, it gets executed. Other than that, `script` can be a
Python file to be executed.
- The pre-defined scripts are now in the folder `src/mmag/scripts`,
while `src/mmag` contains the main package functions.
- Functions were moved around the files in a more "organic" way (in my
opinion)